Made requested modifications to plugin
Added proper plugin headers, saved defaults on admin init, added
internationalization to string, and added filters for subject and
message.  The framework has also been changed to allow for a single
option stored as an array, but since options are still stored in
individual elements due to the use of register_setting, this change has
only been approximated.