Added 'from' filter for alert emails
Added an nbcs-moderation-queue-from filter which allows you to specify
an email address other than the blog admin email for the moderation
queue alerts.  If an invalid address is specified, no From: header is
used at all.  Also, the address is now aliased as "Moderation Queue
Alert" regardless of the specified address.